uint32 now

+++ contracts/PgaDfs.sol
@@ -30,7 +30,7 @@ contract PgaDfs is usingOraclize {
     // MAX of 8 pga tour ids...
     // you can play less than 8 guys if you want!
     // e.g. [web3.fromAscii("21059"), web3.fromAscii("94320"), web3.fromAscii("85933")]
-    int32[8] golferIds;
+    uint16[8] golferIds;
   }
 
   struct Contest {
@@ -139,9 +139,9 @@ contract PgaDfs is usingOraclize {
 
 
   // slate id ==> golfer ids
-  mapping (bytes12 => int32[]) slateIdToGolferIds;
+  mapping (bytes12 => uint16[]) slateIdToGolferIds;
   // slate id ==> pga tour id ==> golfer data (salary, scores, etc.)
-  mapping (bytes12 => mapping(int32 => SlateGolfer)) slateIdToSlateGolfers;
+  mapping (bytes12 => mapping(uint16 => SlateGolfer)) slateIdToSlateGolfers;
   mapping (bytes12 => address[]) slateIdToEnteredAddresses;
   mapping (bytes12 => mapping (address => Lineup)) slateIdToLineups;
 
@@ -181,7 +181,7 @@ contract PgaDfs is usingOraclize {
       return slateIdToEnteredAddresses[slateId];
   }
 
-  function getCurrentSlateLineupForAddress(address address_) public view returns (bytes32, int32[8]) {
+  function getCurrentSlateLineupForAddress(address address_) public view returns (bytes32, uint16[8]) {
       Lineup memory theLineup = slateIdToLineups[slateId][address_];
       return (theLineup.golferIdsHash, theLineup.golferIds);
   }
@@ -190,7 +190,7 @@ contract PgaDfs is usingOraclize {
     return slateId;
   }
 
-  function getSalary(int32 pgaId) public view returns (int8) {
+  function getSalary(uint16 pgaId) public view returns (int8) {
     return slateIdToSlateGolfers[slateId][pgaId].salary;
   }
 
@@ -201,7 +201,7 @@ contract PgaDfs is usingOraclize {
   function revealLineup(string golferIdsColonDelimited, string revealKey) public returns (int16) {
     require(slateIdToLineups[slateId][msg.sender].golferIdsHash == keccak256(strConcat(golferIdsColonDelimited, "|", revealKey)));
 
-    var golferIds = new int32[](8);
+    var golferIds = new uint16[](8);
 
     var golferIdsSlice = golferIdsColonDelimited.toSlice();
     var delimiter = ":".toSlice();
@@ -217,7 +217,7 @@ contract PgaDfs is usingOraclize {
       // string  --> left padded bytes32
       // "29725" --> "0x3239373235000000000000000000000000000000000000000000000000000000"
       // replicate this with web3.fromAscii('29725')
-      int32 golferId = int32(parseInt(golferIdsSlice.split(delimiter).toString()));
+      uint16 golferId = uint16(parseInt(golferIdsSlice.split(delimiter).toString()));
 
       for (uint8 jj = 0; jj < ii; jj++) {
         require(golferId != golferIds[jj]);
@@ -293,7 +293,7 @@ contract PgaDfs is usingOraclize {
       for (uint8 ii = 0; ii < playerCount; ii++) {
           var playerColonSalary = compressedSalariesSlice.split(playerDelimiter);
 
-          int32 pgaPlayerId = int32(parseInt(playerColonSalary.split(salaryDelimiter).toString()));
+          uint16 pgaPlayerId = uint16(parseInt(playerColonSalary.split(salaryDelimiter).toString()));
 
           int8 salary = int8(parseInt(playerColonSalary.toString()));
 
@@ -315,7 +315,7 @@ contract PgaDfs is usingOraclize {
 
     for (uint16 i = 0; i < playerScoreSlices.length; i++) {
       playerScoreSlices[i] = compressedScoresSlice.split(playerDelimiter);
-      int32 pgaPlayerId = int32(parseInt(playerScoreSlices[i].split(":".toSlice()).toString()));
+      uint16 pgaPlayerId = uint16(parseInt(playerScoreSlices[i].split(":".toSlice()).toString()));
       uint roundSlices = playerScoreSlices[i].count("-".toSlice()) + 1;
       for (uint rd = 0; rd < roundSlices; rd++) {
         int8 rdScore = int8(parseInt(playerScoreSlices[i].split("-".toSlice()).toString()));
@@ -334,13 +334,13 @@ contract PgaDfs is usingOraclize {
     require(contest.live);
     require(!contest.slateIdToPayoutsSet[slateId]);
 
-    int32 totalEntries = int32(contest.slateIdToEntries[slateId].length);
+    uint16 totalEntries = uint16(contest.slateIdToEntries[slateId].length);
     int totalPoints = 0;
 
     // calculate the average score in the contest
     for (uint8 ii = 0; ii < totalEntries; ii++) {
       address entry = contest.slateIdToEntries[slateId][ii];
-      int32[8] memory entryPgaIds = slateIdToLineups[slateId][entry].golferIds;
+      uint16[8] memory entryPgaIds = slateIdToLineups[slateId][entry].golferIds;
       for (uint8 g = 0; g < entryPgaIds.length; g++) {
         contest.slateIdToAddressScores[slateId][entry] += slateIdToSlateGolfers[slateId][entryPgaIds[g]].points;
       }
